Fixed bug, added units, and updated length interface style
- Fixed a bug that caused incorrect conversion in the length converter.
- Added support for additional units like inches, miles, and nautical miles.
- Updated the styling of the length converter interface to improve user experience.

+++ Length.cs
@@ -19,7 +19,7 @@ namespace CalculatorGUI
 
         private void btn_toconvert_Click(object sender, EventArgs e)
         {
-            double result = LengthConvertor.Convert(int.Parse(txbox_from.Text), combo_form.Text, combo_to.Text);
+            double result = LengthConvertor.Convert(Double.Parse(txbox_from.Text), combo_form.Text, combo_to.Text);
             txbox_to.Text = result.ToString();
         }
